recovery: Add exec support for OpenWrt routers
Adds a new router controller for OpenWrt routers and adds new
wifirouter execs that adds support for repairing OpenWrt routers.
This includes support for updating OpenWrt router images by checking
build info on the device, comparing that with the expectations of the
config in GCS, and updating routers with images from GCS when needed.

BUG=b:282836912
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...
TEST=successfully ran an updated wifi router recovery plan on a testbed
with Gale routers and confirmed that only the Gale-specific actions
were followed
TEST=successfully ran wifi router recovery plan on a testbed with
an openwrt router and confirmed that the WifiRouterHost object printed
after the local recovery completed was as expected
TEST=successfully ran wifi router recovery plan on a testbed with an
openwrt with an outdated image, and confirmed that the repair plan was
able to update the router image based on GCS configs

Change-Id: I38f73357b053ed21333bce06827676db1518a094