recovery: Add support for using custom ssh usernames
Execs run ssh commands on hosts always as the root user. Not all
devices we plan to add support for in the repair plan execs use
root as the ssh user (e.g. AsusWrt wifi routers). This CL adds
support for using a different user as an option in the ssh run
method.

BUG=b:282836502
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...

Change-Id: I49729f0b68c5d28b6ab41b2a1571bb348ea6036b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4558995
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#57491}