[test-resources] Cache file summaries in the directory rows
The roll ups for directories need to be quick but the api only
will query for a single node's direct ancestors. To keep this fast
and roll up only for filtered files we need to cache the summaries
for those files that affect a directory.

Additionally this CL will:
- Add descriptions. These will appear in the bigquery console
- Add a flag to control the data set for debugging
- Add parameters to the create_tables script for creating the tables
- Fetch based on the filter

Bug: 1435578
Change-Id: Ib557203745e4593f58e7fdcdddcc952e251819c4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4605845
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Riley Wong <rgw@google.com>
Cr-Commit-Position: refs/heads/main@{#57567}