vm-leaser: implement GetZoneSubnet to specify subnet for VM Leaser
Because some networks are in subnet mode, they require VM Leaser to pass
a subnet. Other networks that are not in subnet mode will error when a
subnet is passed. Using SubnetModeNetworkEnabled field, VM Leaser will
determine when to pass a subnet value. The subnet value is based on the
selected zone.

BUG=b/283819096
TEST=make test; make dev; manual grpcurl testing

Change-Id: If71a196a85800cc5b21968c6e622120746320779
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4638254
Reviewed-by: Deniz Kara <denizkara@google.com>
Commit-Queue: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/main@{#57696}