Remove yulanlin@google.com from go/src/infra/appengine/rubber-stamper/OWNERS.
This suggested change is automatically generated based on group memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot will abandon it. Vote CR +1/2 to approve this change.
See the owner's recent activity for context: https://chromium-review.googlesource.com/q/yulanlin@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I056ae443f206a2c1be284ffdceff63972f191866