[smart-scheduling] Add normalize-servo-name function to heuristics
We need this function to implement the RPC that gets the DUTs associated
to a particular labstation efficiently. Since this involves manipulating
hostnames based on a well-established naming convention, put it in the
heuristics library so we perform this heuristic option the same way every
time we need to do it.

BUG=b:268078240
TEST=make test

Change-Id: Id9e9fccb9f004a1389df76986cb441661390568b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4642161
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#57709}