[Builder health updater] Continue when builders no longer exist
When a builder is deleted within 7 days, its health data is available but we cannot set the health of the builder. This causes the job to fail, even though the error is not fatal.

This CL will make the builder health updater job continue. The error will still be logged in the build summary, but the job will not fail.


Before:
$ luci-auth context -- go run . generate_indicators -date=2023-01-01

[I2023-06-26T22:28:14.304143Z 798751 0 errors.go:128] set status: SUCCESS                         {"build.step":"RPC Buildbucket"}
[E2023-06-26T22:28:14.304235Z 798751 0 main.go:152] Error in Run: RPC buildbucket: 13 set builder health requests failed
exit status 1

After:
$ luci-auth context -- go run . generate_indicators -date=2023-01-01

[I2023-06-26T21:47:42.664511Z 771382 0 errors.go:128] set status: SUCCESS
  with SummaryMarkdown:
13 set builder health requests failed {"build.step":"RPC Buildbucket"}


Change-Id: I2726799337898289f066ed8d0b8b7bb41b8f2c73
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4647169
Reviewed-by: Gregory Guterman <guterman@google.com>
Commit-Queue: Trung Nguyen <tne@google.com>
Cr-Commit-Position: refs/heads/main@{#57788}