recovery: Add verbose logs for cellular DUTs
This enables verbose system network logs for DUTs in the cellular pool
and includes them in the repair task outputs in order to generate more
insight into the hardware state of cellular DUTs during the repair
process (b/285637596).

Note: Logs are enabled at the beginning of the process even though they
aren't exported until the audit step to identify if other audit steps
may interfere with the modem.

BUG=b:289109647
TEST=run local, unittests

Change-Id: I42b352de7530d63ff22c00bdb9c0d2d29b07bbb1