[cloudbuildhelper] Resolve `${contextdir}/Dockerfile` if present.
If YAML manifest has `Dockerfile` field, the dockerfile specified
there is "resolved" (tags in it are replaced with sha256 digests)
before being fed to the builder. This is essential for triggering
rebuilds when an upstream tag is "moved": we use hash of the
"resolved" context directory to see if anything changes.

This didn't work (Dockerfile was left unresolved) if the YAML
had only `Contextdir` set, but not `Dockerfile`. Fix that.

R=iannucci@chromium.org

Change-Id: I9f05e53eb4ff2fe09b8c1cf6fc9a09e4d3b05b51
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4660315
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#57950}