skylab: Refactor to use WifiRouterFeature enum
Copied WifiRouterFeature enum from config repo to skylab device proto
and refactored it to use this enum rather than the previous
placeholder enum for peripheral_wifi_features. Also adjusted logic to
reflect actual usage in that it is a simple copy of
wifi_router_features, which is resolved during repair rather than
in the export.

BUG=b:282839248
TEST=cd ~/cr/infra/go/src/infra/libs/skylab/inventory && go generate
TEST=cd ~/cr/infra/go/src/infra/unifiedfleet && make test
TEST=cd ~/cr/infra/go/src/infra/libs/skylab/inventory && go test ./...

Change-Id: I95b50872ad1c0fc8a3c349aa2ea0d5a1c6cbc6ff