ufs: Add allowlist for sfp RPCs
Some RPCs don't care which namespace they are called in (one heuristic-
if they are not hitting datastore). If that's the case, we shouldn't
care about limiting partner access to specific namespace.

The alternative is getting CTP to accept configuration to indicate which
namespace it should hit UFS from when calling CheckFleetTestsPolicy.
Because this would be the only use case, and the namespace effectively
doesn't matter for the RPC, that seemed unnecessarily complex so I went
with a UFS-side change.

BUG=b/285605479
Test Plan=make test, manual test with dev workflow

Change-Id: I51e9075c8044d91718de3858f85ca751a99e6c3f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4670330
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Cr-Commit-Position: refs/heads/main@{#58051}