recovery: Use a repair_failed_count of 3 for Flex devices
We have a limited stock of Flex DUTs in the lab and sometimes a DUT that
is `repair_failed` can take upwards of a day to transition to
`needs_manual_repair`.

This change reduces the `repair_failed_count` for
Flex devices to three, which should reduce the amount of time it takes
for a broken DUT to reach `needs_manual_repair`, allowing Flex lab techs
to manually recover and restore Flex DUTs to the fleet faster.

BUG=b:287632459
TEST=build

Change-Id: I8e600d2d1b2526f1281a50ce84b7c5bc26264c2c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4670329
Reviewed-by: Justin Suen <justinsuen@google.com>
Commit-Queue: Joseph Sussman <josephsussman@google.com>
Cr-Commit-Position: refs/heads/main@{#58069}