Remove akashmukherjee@google.com from go/src/infra/isolation/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/akashmukherjee@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I2a912c03825917e106024031f6c22aec362db8f0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4688691
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Suma Kasa <sumakasa@google.com>
Reviewed-by: Suma Kasa <sumakasa@google.com>
Cr-Commit-Position: refs/heads/main@{#58247}