[golangbuild] add support for test sharding
This change is fairly large with the ultimate goal of supporting test
sharding. This change refactors a few parts of golangbuild, adds a bunch
of helper functions, and then implements different "modes" which are
specified by the input properties.

When running in ALL_MODE, this change should be a no-op right now.

Change-Id: Id52de47211751faf0f78b3f386a11f867099fe59
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4678243
Reviewed-by: Heschi Kreinick <heschi@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#58327}