[golangbuild] include update token in ResultDB invocation RPC
I neglected to add the -include-update-token flag in my last change
which causes the RPC to fail. Adding the flag should trivially work (the
update token is obtained from LUCI_CONTEXT), but it may require giving
the service account a new permission; we'll see. For now, this is
definitely necessary, based on
https://source.chromium.org/chromium/infra/infra/+/main:recipes-py/recipe_modules/resultdb/api.py;l=466;drc=e8e477b715774d7c1d01b35835a439ad6d0196c1

Change-Id: Ia3b6a6295e780cea8d8085ff822c54107fccd67a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4701475
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#58352}