[coverage] Export file coverage data from reviver runs
This change also includes making the bucket of an exported
builder configurable (defaulting to previously hard-coded
value of 'ci'). This is because reviver runs are in a
separate 'reviver' bucket.

Also included is a drive-by update of an obsolete docstring.

Bug: b/288244035
Change-Id: I412b9fc7d76d62269f4c5f92bd12b39ad4e906fd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4705764
Reviewed-by: Prakhar Asthana <pasthana@google.com>
Commit-Queue: Prakhar Asthana <pasthana@google.com>
Auto-Submit: Mateusz Zbikowski <zbikowski@google.com>
Cr-Commit-Position: refs/heads/main@{#58381}