mallet: Add support for providing proxies
Adds a new "dev-host-proxy-addresses" CLI flag to mallet that passes
a JSON map of resource names to proxy hostnames to TLW, which in turn
causes TLW to use these proxy hostnames when trying to ssh into the
resource rather than their names directly like normal. This is similar
to "dev-jump-host", but instead of having TLW create the ssh tunnels
from the jump host, it allows users to create them externally however
they want and just pass them.

BUG=None
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...
TEST=cd ~/cr/infra/go/src/infra/cmd/mallet && go clean && go build
TEST=successfully ran mallet locally with and without the new CLI flag
and confirmed via the logs that the provided proxies were being
configured and used as expected when running the wifirouter plan

Change-Id: Ibec2353437f8370d1812b87b9bb420ebbfd28103