siso: flush depfile if deps is not gcc or msvc
for deps=gcc,msvc, depfile is parsed and record
it in deps log and remove depfile.

otherwise, depfile need to be written.
it is recorded in hashfs, but not flushed to the disk,
so incremental build may consider these files don't exist
and hashfs is not clean.

flush depfile if deps is not gcc,msvc

Bug: b/292452031
Change-Id: Ib64cd2084f2c1fcccbc134a331e26ebc7dc30c62
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4705422
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Reviewed-by: Richard Wang <richardwa@google.com>
Commit-Queue: Richard Wang <richardwa@google.com>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#58396}