[golangbuild] simplify dist test invocation
The cmd/dist command has a test subcommand whose default behavior is to
run all tests in the main Go repo. Its CLI has two ways of selecting a
subset of tests to run: either via the -run flag which takes a (possibly
negated) regexp to match, or by explicitly listing test names to run if
the caller happens to already know the exact tests it wants to run.

The latter is already the case, since we first call dist test -list to
discover all tests in the main Go repo. So use the test name arguments
instead of the equivalent -run regexp value.

cmd/dist happens to have a small optimization where it doesn't need
to do as much work to discover tests if it's given an explicit list
of tests that are all normal Go packages (i.e., without a ":variant"
suffix), and we will begin to benefit from that as already done in
the previous cmd/coordinator codebase.

That said, my original motivation here is to allow the unshare command
wrapping to work with sharding by eliminating the need to deal with
passing special characters like '$' in the command-line invocation.

Change-Id: I94c7a9c0c4eac72edb05887114a412c431ca84ff