cr-rev: support redirect for diff between two git numberer numbers
This functionality is currently built in to test-results.appspot.com,
however it is the last piece of functionality there and we want to turn
down the server.

The only known client of this is Sheriff-o-Matic

Note that there is a slight change to the existing single numberer style
links in that it will find all commits that match and then choose the
best one, in the same style as the full git hash matches.  Please let me
know if this is undesirable behavior and I can remove it.

Change-Id: Id3b821a7d4ba3af3eac52f59e39538c2bddacfb0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4719829
Reviewed-by: Josip Sokcevic <sokcevic@chromium.org>
Commit-Queue: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#58479}