cros_vm_provision_executor: Don't crash on cleanup if no lease.
vmProvisionReleaseCommandExecution could be executed both as a result of
normal execution, and as one of cleanup actions. If it's part of cleanup
there are no guarantees that the VM Leaser was able to lease the DUT
successfully. Currently we crash in this case, and this CL is trying to
fix this by checking for nil pointers.

BUG=b:292436930
TEST=unittests

Change-Id: I8aa73df17fd04d8bc37c67cbdd2602dd7c8c1e06