[infra] fix for less-common operating systems
Use the unix build tag in various places. Update to the newer xattr
package, and switch to the "aix" branch of fsnotify for AIX support.
It's a little out of date, but not too bad, and the author seems
receptive to rebasing it.

Change-Id: Id61bde8435e9abab1c67bd2eee70f7479da25381
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4722679
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#58557}