[golangbuild] make the no-network check fail closed
The no-network v2 attempt to switch to --map-current-user has proven to
need more permissions and more work. I've decided to leave further work
on improving the check for later, after we tackle root builders.
Stay with --map-root-user since it works well enough for today's needs.

Previously golangbuild would automatically opt-in to using the check
when its system requirements were met and it was appropriate, but that
can fail quietly without anyone noticing via logs. Make the main.star
configuration responsible for turning on the check instead, and have
golangbuild fail the build if it can't do what's required of it.

The check is controlled by a new input property now and ready to be
used, so clean up the soon-to-be unused experiment names.

Change-Id: I7fe243bc9a25ca4dc9ab9021fe6c3703a9536580
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4727455
Reviewed-by: Heschi Kreinick <heschi@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#58564}