[cbi-stats] Improve test case counts
Tests that report ask skipped should not be added to the test count
as they don't use the same resources.

The distinct test counts per day generally follows the max tests. This
leads to a CQ attempt that adds a lot of tests increasing the metric
before it lands

Change-Id: I18b404909773ba9f624b6e957e2444608596d14c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4717943
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#58579}