Add field for shadow properties file to chromium bootstrapper.
This adds a field to the bootstrapper's $bootstrap/properties property
that identifies the shadow properties file to use for the builder. This
only adds the field without changing any functionality. This will allow
the field to be set by the builders without causing failures due to
unknown fields so that the functionality change can be tested.

Bug: 1468653
Change-Id: I7d08e1a2a4e8aed187332e02fafbd97a74a54f95
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4732424
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#58580}