[WIB] Fix build summary too long issues
Bug: b/293967631

Fixes the issue with build summary crash. The final build summary is
truncated to 3997 bytes and is terminated with `...` if the summary
goes beyond 4000 bytes. This should in practise only terminate the
did not build list (which is not important info IMO). The complete
build summary is included in execute step.

Change-Id: Idf0a0251fc5e2d5150ee5b12c4b948ec300ff700
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4739673
Commit-Queue: Adam Todd <actodd@google.com>
Reviewed-by: Adam Todd <actodd@google.com>
Cr-Commit-Position: refs/heads/main@{#58610}