recovery: Add router features to TLW
Replaces the existing placeholder router features enum and its usage
with the shared lab WifiRouterFeature enum. Also adds a wifi router
feature list for the overall testbed that will only contain features
common between all routers in the testbed. The WifiRouterHost.board
field was deprecated as it's Gale-specific and is not used by
anything. The WifiRouterHost.device_type field has been added to
track the type of router a device is, which will be used to adjust
exec logic to match the device and will be used in config conditions.

For now, these changes are just on the TLW end so that we can create
the new execs for different types of routers. These changes will be
added to UFS in a later CL.

BUG=b:282836502
TESt=cd ~/cr/infra/go/src/infra/cros/recovery/tlw && go generate
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...
TEST=successfully ran wifi router recovery plan on a testbed with gale
routers

Change-Id: I6eb83d615115c2f576a80d3ac574f8deda6b1837
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4621421
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#58616}