recovery: Add ssh runner interface for routers
Adds a Runner interface in a new ssh package that will be used to
govern ssh command execution. Abstract out TLW and HostAccess
usage from building ssh commands and receiving their results. Is
compatibly with the current TLW runner result and implementations
can use TLW as the runner behind the scenes.

In the future the idea is to generalize this for ssh commands on
any host, but due to time constraints of dependent features this
will be limited for router controllers for now. It can be refactored
for general host access later.

BUG=b:282836502
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean \
&& go build && go test ./...

Change-Id: I1bd592ffc9f385fbc55d92c8341e42f9d6ab1e75
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4671094
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#58618}