recovery: Refactor wifi router repair execs
This refactors the wifi router repair execs to prepare for supporting
multiple types of routers. The same repair steps are performed for
Gales, but they are now reorganized. Some new repair execs have been
added as well, including a device reboot and collection of wifi
features and device type. These fields will be saved to UFS in a
later CL.

With this new design, each type of router is managed by a different
controller implementation. Identification of the router occurs in an
action prior to any device-specific actions and execs so that the execs
will know how to properly interact with the router host when they are
called. Since Gale routers are simple and only one model is supported,
its controller is mostly made up of static data. Other controllers added
in later CLs will be more complex.

BUG=b:282836502
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...
TEST=successfully ran current wifi router recovery plan on a testbed
with gale routers
TEST=successfully ran updated wifi router recovery plan on a testbed
with gale routers and confirmed that the WifiRouterHost object printed
after the local recovery completed was expected with the new fields
populated and the gale reboots successfully

Change-Id: If40402bd0dbd94c436f306fa0da1b9f3444396ff
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4558996
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#58628}