recovery: Refactor wifi router repair config
Updates the router repair configuration to use the new router execs
and structure of separating actions by device type. Technically, this
performs the same repair steps as before for Gales, with the addition
of a reboot action.

BUG=b:282836502
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...
TEST=successfully ran updated wifi router recovery plan on a testbed
with gale routers and confirmed that the WifiRouterHost object printed
after the local recovery completed was expected with the new fields
populated and the gale reboots successfully

Change-Id: I2cc386038347db31a6d88b8e920161a45f50f180
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4621422
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#58629}