ufs: Add router features and type to UFS
Adds WifiRouterFeature fields to Wifi and WifiRouter in UFS. Also
adds a WifiRouterDeviceType field to WifiRouter.

There are placeholder feature fields with local Feature enums that
these new fields are meant to replace, and they will be removed in
later CLs after usage has been moved to the new fields.

BUG=b:282840450
TEST=cd ~/cr/infra/go/src/infra/unifiedfleet && make gen && make test

Change-Id: I5641af5730ae34f5af771f7f84c10b3184fc5c44