[RotationProxy] Add ExpiryAt field
This will add the ExpiryAt field when we receive new updates.

TTL will be enabled for the ExpiryAt field.

It is possibly that the already deleted rotation won't be automatically deleted by TTL (because they don't have the expiry_at field set). In this case, I will manually delete them (there are only 95 rotations in datastore in total, so it should be ok).

Change-Id: If4f1d87060b9f804868e90e74dfc56ca442f58b8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4741272
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/main@{#58645}