Add support for shadow properties to chromium bootstrapper.
In order to allow led real builds for builders in the ci bucket, the
builders are configured to switch to using the try pool and service
account for led builds. This in turn requires switching the reclient
instance since the try service account does not and should not have
access to the trusted reclient instance. Support was added for shadow
properties, but unfortunately, the bootstrapper overwrites them when it
loads the properties from the properties file.

To address this issue, this change adds support for a shadow properties
file to the bootstrapper. If the shadow_properties_file field is set in
the $bootstrap/properties property then when the bootstrapper runs on a
shadow build, after loading properties from the properties file, it will
update them with properties loaded from the shadow properties file.

Bug: 1468653
Change-Id: I15c0291181a0a2daaa4144c6d923b52dc42aebd6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4730320
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#58650}