[golangbuild] begin support for misc-compile presubmit builders
The previous Go build system had a property that each Go CL is always
tested with all Go ports to at least some degree. There are some ports
where we do not have sufficient speed and capacity to be able to meet
the timing requirement for presubmit testing, so for those we fall back
to doing faster compile-only smoke testing that can be done on any
fast and available machine via cross-compilation.

Since only compiling tests is much faster than also running said tests,
but the overhead of starting and ending a build is fairly constant,
this change adds "ports" as a possible other dimension that can be
sharded against. The MiscPorts input property enables this mode.
(It's a single opinionated boolean rather than something more flexible,
since it only has misc-compile builder type as its intended user now.)

The misc-compile trybots end up being nothing more than a combination
of the CompileOnly property together with the MiscPorts property.

Notably port selection is performed as part of test mode because it
can fail due to a bad code change, and since we prefer to run it in
the same security boundary where we are already willing to execute
arbitrary test code even before it is fully reviewed and submitted.

For go.dev/issue/61698.

Change-Id: I5f716177bfe966a9755913f46b6bb226b71bf18b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4739675
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#58690}