recovery: Add WgetURL to ssh command utils
Adds an ssh command util function for running the "wget" command on
remote hosts. In addition to stdout and stderr, the HTTP error code
may be returned when wget returns a non-zero exit code. This HTTP
error code is parsed from the stderr. Unit tests have been included
as well to validate the handling of wget errors.

BUG=b:282836912
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...

Change-Id: Ifb2839054405e51e071a4a6999fec6617716e763