recovery: Add support for fetching router config
Adds support to the OpenWrt router controller for downloading, parsing,
and validating the global router controller config file from GCS
through the use of the cache server and router. The unmarshalled
config is stored in the router controller state for later reference
in future execs. A new exec has been added to trigger this process.

BUG=b:282836912
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...
TEST=successfully ran recovery on an OpenWrt router with a custom
config which called the new exec

Change-Id: I6940428d09ef38ae27ce6c2809cc0d5c261d0e31