recovery: Add OpenWrt router image updating
Adds support to the OpenWrt router controller for updating a router
to the expected OpenWrt OS image for the device, as well as a new exec
to trigger this action.

All ChromeOS OpenWrt image binaries are stored in archives in GCS, and
cataloged in the global config file. The respective archive for the
expected image is downloaded to the router from GCS through the cache
server. The image binary is then extracted from the archive on the
router. The OpenWrt sysupgrade process is then started with the new
image binary. After a time, device reboots and should be running the
new image. The build info file is re-retrieved from the device (which
would be the build info file of the expected image) and its ImageUUID
is checked to make sure it matches the ImageUUID of the expected image.
No cleanup of the archive or binary files are needed, as they are
downloaded to /tmp on the router, which is wiped on router reboot.

BUG=b:282836912
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...
TEST=successfully ran recovery with an OpenWrt router with an outdated
image with a custom config that ran these new OpenWrt execs and
validated that the router was successfully updated to the expected
image

Change-Id: Ia1d016b01b95467cb5bb98a2023e93ec52358ba3