recovery: Add basic OpenWrt router controller
Adds a new router controller for OpenWrt routers that retrieves
the model and supported features from the build info file present on
all ChromeOS OpenWrt test routers. The router identification and
controller initialization steps have been updated to support this
new controller type. A new exec has been added to trigger the
retrieval of the build info file.

BUG=b:282836912
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& go test ./...
TEST=successfully ran the repair plan with an openwrt router with a
custom config that calls the new exec and verified the router output
matched the model and supported features for the device

Change-Id: I286da8657a77d9df478cf0878db5d11df3f5844b