[golangbuild] handle multiple ports in fetchSubrepoAndRunTests
In practice, this will only be used for misc-compile builders at this
time. It's actually way simpler than I anticipated since most of the
code in fetchSubrepoAndRunTests is not at all port-specific. Phew.

Later on it might be worthwhile to add a build step to group all nested
modules tested for each port, but not if there's only 1 port as is the
most common case.

For go.dev/issue/61698.

Change-Id: If6f8f90b43b7632d8619f36f360cb0e17443811c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4749182
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#58704}