[golangbuild] simplify Port struct
If go tool dist list is working as intended, it should never return
a port whose GOOS or GOARCH value is an empty string. Check for that
condition right away in goDistList, and report it as a fatal error.

This lets me simplify the Port struct by dropping its 'explicit' field.
The next CL adds some Port comparisons, and not having that field makes
them easier to write.

Also simplify addPortEnv to be function since it doesn't seem to need
anything from buildSpec at this time. It could also be a method of Port,
but it seems nicer to leave it closer to buildSpec.setEnv as it is.

For go.dev/issue/61698.

Change-Id: I77fdb5bc5fe17290b546a960d01c62cbf7c7d5e2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4752754
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#58708}