Retrieve Chrome and Chromium builders and push them to BigQuery
cr-buildbucket.chrome.builds only shows active builders. We are
interested also in inactive builders (i.e. the builders that are still
configured in LUCI but has not been run), as these builders are likely
low-value and can be removed.

The app will do the following:
1. Retrieve the list of Chrome and Chromium builders
2. Push the list to BigQuery

Test:
Launch
$ luci-auth context -- go run . generate

Then run the following query

SELECT
  COUNT(*)
FROM
  `cr-builder-health-indicators.builder_value.builders`;

which returns 2350

Bug: 1469934
Change-Id: Id46b4ad2a6f4ddbf646c09fae55d19e561357d9e