skylab: Add converters/reverters for new wifi labels
Updates the peripherals dimension converters and reverters to support
the new wifi labels, label-wifi_router_features and
label-wifi_router_models. The support for the existing
label-peripheral_wifi_features label is removed at the same time, since
it was a placeholder field not used by anything and is being
replaced with label-wifi_router_features.

The label-wifi_router_features label parsing allows for unknown enum
values to persist as long as they are ints, since the skylab enum copy
may be out of sync with the config enum, which governs the UFS data
in addition to the source of the UFS data. This should prevent delays
in usage of new feature additions to the enum while skylab updates are
pending.

BUG=b:282839248
TEST=cd ~/cr/infra/go/src/infra/libs/skylab/inventory && go test ./...

Change-Id: If8edcfb1487e4afdf0886223849bba0cb149f3bc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4754320
Reviewed-by: Gowri Denduluri <gowriden@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#58823}