[test-health] Fix AutoComplete Component buggy scrollbar
The fix including removing the vertical scrollbar from
the AutoComplete component. Also repositions the AutoComplete
dropdown button so that the horizontal scrollbar does not appear
when it is not in focus. Adds a fix to a bug where we were sending
[''] as the component rather than [].

Need to fix - Adding Autocomplete component creates a vertical
scrollbar on the browser

Bug: 1435578
Change-Id: Ic59a31706ce1abe39882564fafb79297361044f3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4775123
Auto-Submit: Riley Wong <rgw@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Gary Tong <gatong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#58938}