[test-health] Add integration tests
This integration tests should create a test dataset, create fake
tables for the source data (eg rdb try results) using the real table
schema to ensure the queries that update our rollups are valid with
the real, up-to-date schema. The update endpoint is then run twice
to simulate the multiple update calls per day. A check to ensure no
duplicate rows exist in the roll ups is the only check on the rollup
data added for now. The FetchMetrics endpoint is then tested with
this rollup to check the num runs is correct for the variant and
rolled up test id.

There is a problem with deleting the old dataset where data still in
the streaming buffer will cause the table to not get deleted. This is
going to require a more creative solution since we cannot manually
flush the table which can take up to 90 minutes to flush on its own

Some follow-up checks will be added to verify other metrics and other
rollups/fetch queries. We'll also want to add some checks for features
like variant sort as well as test id sort work as expected

Bug: 1435578
Change-Id: Ia0d5a9ffdcc8afd9da10652ffa8c4190040e64e0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4734691
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#58939}