[rts-suite-analysis] Enable rts-suite-analysis to run for longer times
This change will make the fetch rejections complete. The query is very
ineffective for longer than 1 month. To mitigate this I've also added
an -append flag to the fetch commands. This will allow for multiple
calls to fetch to be aggregated. The tradeoff being that flakes
occurring between these fetches won't be deflaked. That should be so
rare that it's not a concern though.

Bug: 1442022
Change-Id: I184eb56ddadb5e7265e70fe54093d1362b04452f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4763068
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#58952}