Make patch failures in chromium bootstrapper regular failures.
Currently, any error not from the triggered recipe gets treated as an
infra failure, but a patch failing to apply doesn't represent an issue
with the infrastructure, so it should be considered a regular failure as
it is in recipes. This changes it so that the PatchRejected tag on an
error will cause the build's status to be set to FAILURE and the summary
set with a message similar to that provided by bot_update.

Change-Id: I8a83e722381c7bf7d26fb341f03b88ffa5cb58be
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4787031
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>
Cr-Commit-Position: refs/heads/main@{#59043}