[go] Update to go1.21.0.
Use GOTOOLCHAIN=local for now (which means Go will never fetch
a different version of itself on its own), just to keep the diff
small and builds less chaotic (think of it as a variant of
"one version policy"). If we find a compelling reason to use
dynamic toolchains, this decision can be reconsidered.

R=iannucci@chromium.org

Change-Id: I0400869709b3d95bfd4b9e8b294f62a21422c597
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4794652
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#59111}