[test-health] Improve table updating
The update has been taking an increasingly long time and is scanning
more data as time goes on. This appears to be because the target does
not have a partition window and is trying to merge over the whole
table.

Also remove the partition expiration since this is set by terraform:
https://source.corp.google.com/piper///depot/google3/configs/cloud/gong/services/chrome_cloud/chrome_browser_infra/chrome_test_health/envs/staging/main.tf;l=44

Bug: 1435578
Change-Id: Iebe77d94ae6256c8e286f1f0f11caf6a962efd87
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4798936
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#59129}