[golangbuild] use the right GOOS/GOARCH from the coordinator builder
Currently coordinator mode builds use hostGOOS and hostGOARCH to check
for a prebuilt Go which is wrong, and ends up with broken builds. Add an
input property for coordinator builds to accept the correct GOOS/GOARCH
explicitly, and pass that through to the prebuilt Go key.

Change-Id: I44f8c7dd5c8c72e3553277657d49341e48bc847d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4799393
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Cr-Commit-Position: refs/heads/main@{#59153}