[golangbuild] always 'go mod download' ahead of time, detect infra fails
The 'go mod download' invocations may fail because the code change
being tested happens to be introduce a real problem to the go command.
This happens almost never.

Another reason it may fail is because the code change introduces a bad
change to the content of the go.mod file. For example, adding a module
version that doesn't exist or has a fatal error in its own go.mod file.
This happens rarely.

Another reason it may fail is due to a temporary network error, because
the common assumption on the internet is that it's cheaper to let the
client try again. This happens sometimes. When it does happen, it's more
useful to annotate the build failure as an infra failure. (Possibly we
can even add internal retries, but I'm leaving that out of scope now.)

Change-Id: I04202313cb27bf5497de9fdee3338b9fc49f4d29
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4806505
Reviewed-by: Heschi Kreinick <heschi@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#59227}