[golangbuild] allow empty patches
Running git cherry-pick on an empty patch fails unless the --allow-empty
flag is set. A patch will be empty if the CL is empty. Such CLs are not
very useful (other than when testing the CI system itself), but failing
with an infrastructure error is also suboptimal.

The previous x/build system works on empty CLs and has proven it's good
enough to rely on code review to prevent empty CLs from being submitted.

The flag is present in Git 2.17.3, so shouldn't get in the way of using
older Git versions as part of go.dev/issue/26746.

Change-Id: I4abab42201dbf5933eb05665e79d407e894c5f48
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4811495
Auto-Submit: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Heschi Kreinick <heschi@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Cr-Commit-Position: refs/heads/main@{#59232}